Partial revert of issue 106
Fixes #114, now if the container is logging output without a trailing
newline, then the output will be written anyway after a 100ms delay as
before 106. However, in the Flush() method of ioproxy now checking
for the number of bytes and when 0 (which occurs after the connection
is closed because of a defer Flush() when using ioproxy), no bytes are
flushed.